Update patch set 1
Patch Set 1:

FYI - this is the first of two commits. This is the client code, next review will have the OVP testsuite. 

Waiting to merge it in until the verify job is working, there was a typo in the previous commit so I've made another patch there first. But if you have a chance to briefly review prior, that'd be great.

Patch-set: 1